refactor(participants): 패키지 분리와 스트림 문법으로 정리, 정적 팩토리 메서드 정상화
