+++ src/main/java/blackjack/domain/participant/Players.java
@@ -1,8 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
 import blackjack.domain.card.Deck;
-import blackjack.domain.participant.Dealer;
-import blackjack.domain.participant.Player;
 import java.util.List;
 
 public class Players {
@@ -10,19 +8,25 @@ public class Players {
     private final List<Player> playerList;
     
     private Players(List<Player> playerList) {
+        validate(playerList);
         this.playerList = playerList;
     }
     
-    public static Players from(List<Player> players) {
+    public static Players fromNames(List<String> playerNames) {
+        List<Player> players = playerNames.stream()
+                .map(Player::new)
+                .toList();
         return new Players(players);
     }
     
-    public static Players makePlayers(List<String> names) {
-        List<Player> result = names.stream()
-                .map(Player::new)
-                .toList();
-        
-        return from(result);
+    private void validate(List<Player> playerList) {
+        long distinctCount = playerList.stream()
+                .map(Player::getNickname)
+                .distinct()
+                .count();
+        if (distinctCount != playerList.size()) {
+            throw new IllegalArgumentException("플레이어 이름은 중복될 수 없습니다.");
+        }
     }
     
     public void distributeCards(Deck deck) {
@@ -38,11 +42,5 @@ public class Players {
                 .filter(Player::isDrawable)
                 .findFirst()
                 .orElse(null);
-    }
-    
-    public List<GameResult> determinePlayersGameResult(Dealer dealer) {
-        return playerList.stream()
-                .map(player -> player.calculateResult(dealer))
-                .toList();
     }
 }
