+++ src/main/java/blackjack/view/InputView.java
@@ -5,31 +5,30 @@ import java.util.List;
 import java.util.Scanner;
 
 public class InputView {
-
-
+    
+    
     private static final String DELIMITER = ",";
-
+    
     private final Scanner scanner = new Scanner(System.in);
-
+    
     public String readLine() {
         return scanner.nextLine();
     }
-
+    
     public List<String> parsePlayerNames() {
         List<String> playerNames = Arrays
                 .stream(readLine().split(DELIMITER))
                 .map(String::trim)
                 .toList();
-
+        
         validateNames(playerNames);
         return playerNames;
     }
-
+    
     private void validateNames(List<String> playerNames) {
         validateName(playerNames);
-        validateDuplicate(playerNames);
     }
-
+    
     private void validateName(List<String> playerNames) {
         playerNames.forEach(name -> {
             if (name.isEmpty()) {
@@ -37,14 +36,7 @@ public class InputView {
             }
         });
     }
-
-    private void validateDuplicate(List<String> playerNames) {
-        long distinct = playerNames.stream().distinct().count();
-        if (distinct != playerNames.size()) {
-            throw new IllegalArgumentException("플레이어 이름은 중복될 수 없습니다.");
-        }
-    }
-
+    
     public boolean getUserCommand() {
         return UserCommand.from(readLine()) == UserCommand.YES;
     }
